projttgs/requirements.txt
# Django Framework
# Using Django 5.1 for Python 3.13 compatibility
Django>=5.1,<6.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# MongoDB - Using mongoengine instead of djongo
# djongo is incompatible with Django 4.2+ and Python 3.13
# mongoengine is actively maintained and works with modern Django/Python
mongoengine>=0.27.0
pymongo>=4.6.0

# Environment variables
python-decouple>=3.8

# Django extensions
django-extensions>=3.2.0

# CORS headers for Next.js frontend
django-cors-headers>=4.3.0

# PDF generation
xhtml2pdf>=0.2.11

# Image processing
Pillow>=12.0.0

# Django dependencies
sqlparse>=0.3.0
pytz>=2021.1
asgiref>=3.2.10
